+++ storage/fs/fs_store.go
@@ -250,20 +250,20 @@ func (store *FSStore) GetObject(checksum string) ([]byte, error) {
 	return data, nil
 }
 
-func (store *FSStore) GetObjectRefCount(checksum string) (uint16, error) {
+func (store *FSStore) GetObjectRefCount(checksum string) (uint64, error) {
 	st, err := os.Stat(store.PathObject(checksum))
 	if err != nil {
 		return 0, err
 	}
-	return st.Sys().(*syscall.Stat_t).Nlink - 1, nil
+	return uint64(st.Sys().(*syscall.Stat_t).Nlink - 1), nil
 }
 
-func (store *FSStore) GetChunkRefCount(checksum string) (uint16, error) {
+func (store *FSStore) GetChunkRefCount(checksum string) (uint64, error) {
 	st, err := os.Stat(store.PathChunk(checksum))
 	if err != nil {
 		return 0, err
 	}
-	return st.Sys().(*syscall.Stat_t).Nlink - 1, nil
+	return uint64(st.Sys().(*syscall.Stat_t).Nlink - 1), nil
 }
 
 func (store *FSStore) PutObject(checksum string, data []byte) error {
