+++ storage/storage.go
@@ -239,7 +239,7 @@ func (store *Store) PutObject(checksum string, data []byte) error {
 	return store.backend.PutObject(checksum, data)
 }
 
-func (store *Store) GetObjectRefCount(checksum string) (uint16, error) {
+func (store *Store) GetObjectRefCount(checksum string) (uint64, error) {
 	t0 := time.Now()
 	defer func() {
 		logger.Profile("storage: GetObjectRefCount(%s): %s", checksum, time.Since(t0))
@@ -271,7 +271,7 @@ func (store *Store) PutChunk(checksum string, data []byte) error {
 	return store.backend.PutChunk(checksum, data)
 }
 
-func (store *Store) GetChunkRefCount(checksum string) (uint16, error) {
+func (store *Store) GetChunkRefCount(checksum string) (uint64, error) {
 	t0 := time.Now()
 	defer func() {
 		logger.Profile("storage: GetChunkRefCount(%s): %s", checksum, time.Since(t0))
