+++ storage/storage_types.go
@@ -46,13 +46,13 @@ type StoreBackend interface {
 	GetObject(checksum string) ([]byte, error)
 	CheckObject(checksum string) (bool, error)
 	PutObject(checksum string, data []byte) error
-	GetObjectRefCount(checksum string) (uint16, error)
+	GetObjectRefCount(checksum string) (uint64, error)
 
 	GetChunks() ([]string, error)
 	GetChunk(checksum string) ([]byte, error)
 	CheckChunk(checksum string) (bool, error)
 	PutChunk(checksum string, data []byte) error
-	GetChunkRefCount(checksum string) (uint16, error)
+	GetChunkRefCount(checksum string) (uint64, error)
 
 	Purge(id string) error
